+++ client/src/Screens/R-egisterScreen.jsx
@@ -1,26 +1,26 @@
 import React from "react";
-import LoginLayout from "../layout/Login-Layout/LoginLayout";
+import LoginLayout from "../layout/LoginLayout/LoginLayout";
 import { Link } from "react-router-dom";
 import { useState } from "react";
 import '../assets/css/login.css'
-import { useNavigate } from 'react-router-dom';
 
+function RegisterScreen() {
 
-function EditProfileScreen( props ) {
-  console.log(props.location);
-  const navigate = useNavigate();
-
-  function goBack(e) {
-    e.preventDefault(); 
-    navigate(-1);
-  }
-
+  const [email, setEmail] = useState();
+  const [password, setPassword] = useState();
   const [displayname, setDisplayname] = useState();
   const [height, setHeight] = useState();
   const [weight, setWeight] = useState();
   const [gender, setGender] = useState();
   const [image, setImage] = useState();
 
+  const handleChangeEmail = (event) => {
+      setEmail(event.target.value);
+      console.log(event.target.value);
+  };
+  const handleChangePassword = (event) => {
+      setPassword(event.target.value);
+  };
   const handleChangeDisplayname = (event) => {
       setDisplayname(event.target.value);
   };
@@ -37,39 +37,67 @@ function EditProfileScreen( props ) {
       setImage(event.target.value);
   };
 
-  // const handleSave = (event) => {
-  //   event.preventDefault(); 
+  const handleSave = (event) => {
+    event.preventDefault(); 
   
-  //   const formData = {
-  //     displayname,
-  //     height,
-  //     weight,
-  //     gender,
-  //     image,
-  //   };
+    const formData = {
+      email,
+      password,
+      displayname,
+      height,
+      weight,
+      gender,
+      image,
+    };
   
-  //   console.log(formData);
+    console.log(formData);
   
-  // };
+  };
 
-  // const handleCancel = () => {
-  //   setDisplayname("");
-  //   setHeight("");
-  //   setWeight("");
-  //   setGender("");
-  //   setImage("");
-  // };
+  const handleCancel = () => {
+    setEmail("");
+    setPassword("");
+    setDisplayname("");
+    setHeight("");
+    setWeight("");
+    setGender("");
+    setImage("");
+  };
+  
 
     return (
       <LoginLayout>
         <form action="" className="form-login form-regis">
 
           <div className="regis-text-top" style={{ marginBottom: "10px" }}>
-            <h2 style={{ fontWeight: "bold" }}>Edit Profile</h2>
+            <h2 style={{ fontWeight: "bold" }}>Register</h2>
           </div>
 
-          <div className="regis-text" style={{ marginBottom: "10px" }}></div>
-          <div className="regis-text" style={{ marginBottom: "10px" }}></div>
+          <div className="regis-text" style={{ marginBottom: "10px" }}>
+            <label htmlFor="email">Email</label>
+            <input
+              type="email"
+              name="email"
+              placeholder=""
+              className="input-regis"
+              style={{ boxShadow: "0px 4px 4px rgba(0, 0, 0, 0.25)" }}
+              onChange={handleChangeEmail}
+            />
+          </div>
+
+          <div className="regis-text" style={{ marginBottom: "10px" }}>
+            <label htmlFor="password" style={{ display: "flex" }}>
+              Password
+            </label>
+            <input
+              type="password"
+              name="password"
+              placeholder=""
+              className="input-regis"
+              style={{ boxShadow: "0px 4px 4px rgba(0, 0, 0, 0.25)" }}
+              onChange={handleChangePassword}
+            />
+          </div>
 
           <div className="regis-text" style={{ marginBottom: "10px" }}>
             <label htmlFor="password" style={{ display: "flex" }}>
@@ -80,11 +108,13 @@ function EditProfileScreen( props ) {
               name="DisplayName"
               placeholder=""
               className="input-regis"
+              style={{ boxShadow: "0px 4px 4px rgba(0, 0, 0, 0.25)" }}
               onChange={handleChangeDisplayname}
             />
           </div>
 
           <div className="hw">
+
             <div className="regis-text" style={{ marginBottom: "10px" }}>
               <label htmlFor="password" style={{ display: "flex" }}>
                 Height
@@ -94,6 +124,7 @@ function EditProfileScreen( props ) {
                 name="Height"
                 placeholder=""
                 className="input-regis"
+                style={{ boxShadow: "0px 4px 4px rgba(0, 0, 0, 0.25)" }}
                 onChange={handleChangeHeight}
               />
             </div>
@@ -107,11 +138,12 @@ function EditProfileScreen( props ) {
                 name="Weight"
                 placeholder=""
                 className="input-regis weight-regis"
+                style={{ boxShadow: "0px 4px 4px rgba(0, 0, 0, 0.25)" }}
                 onChange={handleChangeWeight}
               />
             </div>
-          </div>
 
+          </div>
           <div className="regis-text" style={{ fontSize: "18px" }}>
             <label htmlFor="gender">Gender</label>
             <input type="radio" name="gender" onChange={handleChangeGender} />
@@ -141,20 +173,22 @@ function EditProfileScreen( props ) {
               <button
                 className="btn-save-regis"
                 style={{ marginRight: "26px", fontWeight: "bold" }}
-                onClick={(e) => goBack(e)}
+                onClick={handleSave}
               >
                 Save
               </button>
             </div>
 
             <div>
-              <button
-                className="btn-cancel-regis"
-                style={{ fontWeight: "bold" }}
-                onClick={(e) => goBack(e)}
-              >
-                Cancel
-              </button>
+              <Link to="/login">
+                <button
+                  className="btn-cancel-regis"
+                  style={{ fontWeight: "bold" }}
+                  onClick={handleCancel}
+                >
+                  Cancel
+                </button>
+              </Link>
             </div>
           </div>
 
@@ -163,4 +197,4 @@ function EditProfileScreen( props ) {
     );
 }
 
-export default EditProfileScreen
+export default RegisterScreen
