+++ client/src/main.jsx
@@ -9,13 +9,13 @@ import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import Dashboard from "./layout/DashboardLayout/Dashboard.jsx";
 import ActivityCardScreen from "./Screens/ActivityCardScreen.jsx";
 import ActivityDetailScreen from "./Screens/ActivityDetailScreen.jsx";
-import EditProfileScreen from "./Screens/EditProfileScreen.jsx";
+import EditProfileScreen from "./Screens/E-ditProfileScreen.jsx";
 import GoalSelectionScreen from "./Screens/GoalSelectionScreen.jsx";
 import LandingPage from "./Screens/LandingPage.jsx";
-import LoginScreen from "./Screens/LoginScreen.jsx";
-import NewPassScreen from "./Screens/NewPassScreen.jsx";
-import RegisterScreen from "./Screens/RegisterScreen.jsx";
-import ResetPassScreen from "./Screens/ResetPassScreen.jsx";
+import LoginScreen from "./Screens/L-oginScreen.jsx";
+import NewPassScreen from "./Screens/N-ewPassScreen.jsx";
+import RegisterScreen from "./Screens/R-egisterScreen.jsx";
+import ResetPassScreen from "./Screens/R-esetPassScreen.jsx";
 import SpecificGoalScreen from "./Screens/SpecificGoalScreen.jsx";
 import SuccessScreen from "./Screens/SuccessScreen.jsx";
 import CRUDScreen from "./Screens/CRUDScreen.jsx";
